Time: 15 ms (83.38%), Space: 22.4 MB (73.25%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -17,7 +17,7 @@ public:
         q.push(root);
         while(!q.empty()){
             int sz=q.size();
-            long double f=0;
+            double f=0;
             for(int i=0;i<sz;i++){
                 auto node=q.front();
                 q.pop();
